move bl_romapi to romapi_flash.c for rt1011
